+++ remote_api/python/utils/ecosystem_lib.py
@@ -2,6 +2,7 @@ from utils.emr_lib import EMRLib
 import os
 import argparse
 import datetime
+import boto3
 
 arg_parser = argparse.ArgumentParser()
 
@@ -55,14 +56,25 @@ arg_parser.add_argument(
     '--s3_data_dir',
     help='S3 path to be used as non-ephemeral data directory.')
 arg_parser.add_argument(
-    '--s3_metadata_dir',
-    help='''If it is not empty, it contains the S3 path of saved metadata,
-  which will be restored after the cluster was started. The metadata will not be
+    '--restore_metadata',
+    action='store_true',
+    help='''If it is set, metadata will be reloaded from `s3_data_dir/metadata_backup/`,
+  after the cluster was started. The metadata will not automatically be
   saved when the cluster is shut down.''')
+arg_parser.add_argument(
+    '--s3_metadata_version',
+    help='''If specified, it defines the VERSION part of the metadata backup directory:
+    s3_data_dir/metadata_backup/VERSION The format of this flag (and VERSION) is
+    `YYYYMMddHHmmss` e.g. `20170123164600`. If not specified , the script will use the
+    latest version in `s3_data_dir/metadata_backup/`.''')
 arg_parser.add_argument(
     '--owner',
     default=os.environ['USER'],
     help='''The responsible person for this EMR cluster.''')
+arg_parser.add_argument(
+    '--kite_instance_name',
+    default='ecosystem_test',
+    help='This sets the KITE_INSTANCE environment variable for LynxKite')
 arg_parser.add_argument(
     '--expiry',
     default=(
@@ -93,12 +105,16 @@ class Ecosystem:
         'lynx_version': args.lynx_version,
         'lynx_release_dir': args.lynx_release_dir,
         'log_dir': args.log_dir,
+        'kite_instance_name': args.kite_instance_name,
         's3_data_dir': args.s3_data_dir,
-        's3_metadata_dir': args.s3_metadata_dir,
+        'restore_metadata': args.restore_metadata,
+        's3_metadata_version': args.s3_metadata_version,
+        's3_metadata_dir': '',
     }
     self.cluster = None
     self.instances = []
     self.jdbc_url = ''
+    self.s3_client = None
 
   def launch_cluster(self):
     print('Launching an EMR cluster.')
@@ -108,6 +124,7 @@ class Ecosystem:
         ec2_key_file=conf['ec2_key_file'],
         ec2_key_name=conf['ec2_key_name'],
         region=conf['emr_region'])
+    self.s3_client = lib.s3_client
     self.cluster = lib.create_or_connect_to_emr_cluster(
         name=conf['cluster_name'],
         log_uri=conf['emr_log_uri'],
@@ -140,16 +157,38 @@ class Ecosystem:
         lk_conf['lynx_version'],
         lk_conf['biggraph_releases_dir'])
     self.install_native_dependencies()
+    self.set_s3_metadata_dir(
+        lk_conf['s3_data_dir'],
+        lk_conf['s3_metadata_version'])
     self.config_and_prepare_native(
         lk_conf['s3_data_dir'],
+        lk_conf['kite_instance_name'],
         conf['emr_instance_count'])
     self.config_aws_s3_native()
     self.start_monitoring_on_extra_nodes_native(conf['ec2_key_file'])
     self.start_supervisor_native()
     print('LynxKite ecosystem was started by supervisor.')
 
+  def set_s3_metadata_dir(self, s3_bucket, metadata_version):
+    if s3_bucket and self.lynxkite_config['restore_metadata']:
+      # s3_bucket = 's3://bla/, bucket = 'bla'
+      bucket = s3_bucket.split('/')[2]
+      if not metadata_version:
+        # Gives back the "latest" version from the `s3_bucket/metadata_backup/`.
+        # http://boto3.readthedocs.io/en/latest/reference/services/s3.html#examples
+        paginator = self.s3_client.get_paginator('list_objects')
+        result = paginator.paginate(Bucket=bucket, Prefix='metadata_backup/', Delimiter='/')
+        # Name of the alphabetically last folder without the trailing slash.
+        # If 'metadata_backup' is missing or empty, the following line throws an exception.
+        version = sorted([prefix.get('Prefix')
+                          for prefix in result.search('CommonPrefixes')])[-1][:-1]
+      else:
+        version = metadata_version
+      self.lynxkite_config['s3_metadata_dir'] = 's3://{buc}/{ver}/'.format(
+          buc=bucket,
+          ver=version)
+
   def restore_metadata(self):
-    # TODO versioning metadata
     s3_metadata_dir = self.lynxkite_config['s3_metadata_dir']
     print('Restoring metadata from {dir}...'.format(dir=s3_metadata_dir))
     self.cluster.ssh('''
@@ -159,7 +198,7 @@ class Ecosystem:
       if [ -d metadata/lynxkite ]; then
         mv metadata/lynxkite metadata/lynxkite.$(date "+%Y%m%d_%H%M%S_%3N")
       fi
-      aws s3 sync {dir} metadata/lynxkite/ --quiet
+      aws s3 sync {dir} metadata/lynxkite/ --exclude "*$folder$" --quiet
       supervisorctl start lynxkite
     '''.format(dir=s3_metadata_dir))
     print('Metadata restored.')
@@ -254,7 +293,7 @@ class Ecosystem:
     mysql -uroot -proot -e "GRANT ALL PRIVILEGES ON *.* TO 'root'@'%' IDENTIFIED BY 'root'"
     ''')
 
-  def config_and_prepare_native(self, s3_data_dir, emr_instance_count):
+  def config_and_prepare_native(self, s3_data_dir, kite_instance_name, emr_instance_count):
     hdfs_path = 'hdfs://$HOSTNAME:8020/user/$USER/lynxkite/'
     if s3_data_dir:
       data_dir_config = '''
@@ -277,7 +316,7 @@ class Ecosystem:
       sed -i -n '/# ---- the below lines were added by test_ecosystem.py ----/q;p'  config/central
       cat >>config/central <<'EOF'
 # ---- the below lines were added by test_ecosystem.py ----
-        export KITE_INSTANCE=ecosystem-test
+        export KITE_INSTANCE={kite_instance_name}
         export KITE_MASTER_MEMORY_MB=8000
         export NUM_EXECUTORS={num_executors}
         export EXECUTOR_MEMORY=18g
@@ -299,7 +338,10 @@ EOF
       # TODO: Find a more sane directory.
       sudo mkdir -p /tasks_data
       sudo chmod a+rwx /tasks_data
-    '''.format(num_executors=emr_instance_count - 1, data_dir_config=data_dir_config))
+    '''.format(
+        kite_instance_name=kite_instance_name,
+        num_executors=emr_instance_count - 1,
+        data_dir_config=data_dir_config))
 
   def config_aws_s3_native(self):
     self.cluster.ssh('''