+++ test/com/lynxanalytics/biggraph/frontend_operations/LogisticRegressionTest.scala
@@ -20,10 +20,28 @@ class LogisticRegressionTest extends OperationsTestBase {
       )
     )
     val classification = project.vertexAttributes("classification").runtimeSafeCast[Double]
-    val probability = project.vertexAttributes("classification_probability").runtimeSafeCast[Double]
-    val probabilityMap = probability.rdd.collect.toMap
+    val classificationMap = classification.rdd.collect.toMap
+    val certainty = project.vertexAttributes("classification_certainty").runtimeSafeCast[Double]
+    val certaintyMap = certainty.rdd.collect.toMap
+    val probabilityOf0 =
+      project.vertexAttributes("classification_probability_of_0").runtimeSafeCast[Double]
+    val probabilityOf0Map = probabilityOf0.rdd.collect.toMap
+    val probabilityOf1 =
+      project.vertexAttributes("classification_probability_of_1").runtimeSafeCast[Double]
+    val probabilityOf1Map = probabilityOf1.rdd.collect.toMap
+
     // Example graph age: 0 -> 20.3, 1 -> 18.2, 2 -> 50.3, 3 -> 2.
-    assert(classification.rdd.collect.toMap == Map(0 -> 0.0, 1 -> 0.0, 2 -> 1.0, 3 -> 0.0))
-    assert(probabilityMap(3) > probabilityMap(1) && probabilityMap(1) > probabilityMap(0))
+    assert(classificationMap == Map(0 -> 0.0, 1 -> 0.0, 2 -> 1.0, 3 -> 0.0))
+    assert(certaintyMap(3) > certaintyMap(1) && certaintyMap(1) > certaintyMap(0))
+    classificationMap.map {
+      case (id, cl) => {
+        if (cl == 0) {
+          assert(certaintyMap(id) == probabilityOf0Map(id))
+        } else {
+          assert(certaintyMap(id) == probabilityOf1Map(id))
+        }
+        assert(probabilityOf0Map(id) + probabilityOf1Map(id) == 1)
+      }
+    }
   }
 }