+++ web/test/tests/segmentation-test.js
@@ -58,4 +58,36 @@ module.exports = function(fw) {
       lib.left.expectCurrentProjectIs('test-example');
       lib.right.expectCurrentProjectIsError();
     });
+
+  fw.transitionTest(
+    'test-example project with example graph',
+    'segmentation size reporting - non empty segments',
+    function() {
+      var params = {
+        'name': 'self',
+      };
+      lib.left.runOperation('Copy graph into a segmentation', params);
+      lib.left.openSegmentation('self');
+    },
+    function() {
+      expect(lib.right.getValue('segment-count')).toBe(4);
+      expect(lib.right.getValue('total-segment-size')).toBe(4);
+      expect(lib.right.getValue('total-segment-coverage')).toBe(4);
+    });
+
+  fw.transitionTest(
+    'segmentation size reporting - non empty segments',
+    'segmentation size reporting - has empty segments',
+    function() {
+      var params = {
+        'filterva-income': '*',
+      };
+      lib.left.runOperation('Filter by attributes', params);
+    },
+    function() {
+      expect(lib.right.getValue('segment-count')).toBe(4);
+      expect(lib.right.getValue('total-segment-size')).toBe(2);
+      expect(lib.right.getValue('total-segment-coverage')).toBe(2);
+      expect(lib.right.getValue('non-empty-segment-count')).toBe(2);
+    });
 };